Adding node return for cache call

+++ src/utils/asyncfn.rs
@@ -34,7 +34,7 @@ pub async fn run() {
         open_cli().await;
     });
 
-    let mut num_node: usize = 0;
+    let mut num_node: u8 = 0;
     let mut list: cache::List = cache::run();
     thread::sleep(Duration::from_secs(1));
